[S12P31A607-157] be/fix: user stat api - type match

+++ backend/api/src/main/java/com/haruspeak/api/moment/application/TodayService.java
@@ -158,7 +158,7 @@ public class TodayService {
      * @param date 날짜
      * @return count
      */
-    public int getTodayMomentCount(int userId, LocalDate date) {
+    public long getTodayMomentCount(int userId, LocalDate date) {
         return todayRedisRepository.countByUserAndDate(userId, date);
     }
